# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Talent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff visible to the current user
      responses:
        "200":
          description: Staff members
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffListResponse" }
        "403": { description: Organization member access required }
  /staff/invite:
    post:
      operationId: inviteStaff
      tags: [staff]
      summary: Invite a teammate to the current org (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteStaffRequest" }
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteStaffResponse" }
        "400": { description: Invalid input }
        "403": { description: Owner access required }
        "409": { description: Email already in use }
  /staff/{id}:
    delete:
      operationId: removeStaff
      tags: [staff]
      summary: Remove a teammate (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400": { description: Invalid input }
        "403": { description: Owner access required }
        "404": { description: Member not found }
  /staff/{id}/role:
    patch:
      operationId: updateStaffRole
      tags: [staff]
      summary: Change a teammate's org role (owner / super_admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStaffRoleRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffMemberResponse" }
        "400": { description: Invalid input }
        "403": { description: Owner access required }
        "404": { description: Member not found }
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    StaffMember:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        fullName: { type: string }
        role:
          type: string
          enum: [employer, institution, admin]
        orgRole: { type: string, nullable: true }
        status:
          type: string
          enum: [pending, active, rejected, invited]
        employerId: { type: integer, nullable: true }
        institutionId: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, email, fullName, role, orgRole, status, employerId, institutionId, createdAt]
    StaffListResponse:
      type: object
      properties:
        members:
          type: array
          items: { $ref: "#/components/schemas/StaffMember" }
      required: [members]
    InviteStaffRequest:
      type: object
      properties:
        email: { type: string }
        fullName: { type: string }
        orgRole: { type: string }
      required: [email, fullName, orgRole]
    InviteStaffResponse:
      type: object
      properties:
        member: { $ref: "#/components/schemas/StaffMember" }
        setupUrl:
          type: string
          nullable: true
          description: One-time password setup URL. Only returned when emailSent is false (no email provider configured); otherwise null so the inviter cannot read someone else's token.
        expiresAt: { type: string, format: date-time }
        emailSent:
          type: boolean
          description: True if the setup link was delivered by the email provider; false if email is not configured.
      required: [member, expiresAt, emailSent]
    UpdateStaffRoleRequest:
      type: object
      properties:
        orgRole: { type: string }
      required: [orgRole]
    StaffMemberResponse:
      type: object
      properties:
        member: { $ref: "#/components/schemas/StaffMember" }
      required: [member]
...